+++ changelog.md
@@ -8,6 +8,8 @@ Codion Change Log
 - Database.DATABASE_INIT_SCRIPTS renamed INIT_SCRIPTS.
 - Database.DATABASE_URL renamed URL.
 - ConnectionPoolWrapper, configuration values renamed, DEFAULT prefix removed.
+### is.codion.swing.common.ui
+- FilterTableColumn.setFixedWidth() added.
 ### is.codion.framework.domain
 - ColumnDefinition.get(ResultSet) overloaded without index, fetches value by column name.
 - KeyGenerator.returnGeneratedKeys() renamed generatedKeys().